requirements.txt
flask
tensorflow-cpu
numpy
opencv-python-headless
flask-cors